pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "sentinel"
version = "0.1.0"
description = "Market intelligence & stock prediction platform combining price, volume, and social sentiment signals."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Brant Anderson" }]
keywords = ["finance", "quant", "machine-learning", "sentiment", "stocks", "backtesting"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

dependencies = [
    # Data
    "pandas>=2.1",
    "numpy>=1.26",
    "pyarrow>=15.0",

    # Market data
    "yfinance>=0.2.40",

    # Storage
    "duckdb>=1.0",

    # ML
    "scikit-learn>=1.4",

    # CLI / UX
    "typer>=0.12",
    "rich>=13.7",

    # Config
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# Social ingestion: Reddit (praw) + X/Twitter (tweepy v2) + VADER scorer
social = [
    "praw>=7.7",
    "tweepy>=4.14",
    "vaderSentiment>=3.3.2",
]

# Gradient-boosted models (XGBoost, LightGBM)
ml-extra = [
    "xgboost>=2.0",
    "lightgbm>=4.3",
]

# Finance-tuned transformer sentiment (finBERT via HuggingFace)
transformers = [
    "transformers>=4.40",
    "torch>=2.2",
]

# Experiment tracking (MLflow)
tracking = [
    "mlflow>=2.12",
]

# Feature importance (SHAP)
explain = [
    "shap>=0.45",
]

# Postgres / TimescaleDB storage backend (opt-in; DuckDB is the default)
postgres = [
    "psycopg[binary]>=3.1",
]

# Crypto OHLCV ingestion via CCXT (unified exchange interface)
crypto = [
    "ccxt>=4.3",
]

dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-pyyaml",
]

[project.scripts]
sentinel = "sentinel.cli:app"

[project.urls]
Homepage = "https://github.com/Anderson-Brant/sentinel"
Repository = "https://github.com/Anderson-Brant/sentinel"

[tool.hatch.build.targets.wheel]
packages = ["src/sentinel"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # bugbear
    "UP",  # pyupgrade
    "SIM", # simplify
]
ignore = ["E501"]  # line length handled by formatter

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
filterwarnings = [
    "ignore::DeprecationWarning:yfinance.*",
    "ignore::FutureWarning:yfinance.*",
]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
